refactor: extract ingress ACL enforcement into dedicated acl-enforcement.ts module
